int changed to double in setSalary method

+++ dependencyInjection/src/main/java/com/example/dependencyInjection/model/Developer.java
@@ -38,7 +38,7 @@ public class Developer {
         return salary;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
